Refactor MediaManager state into hook and view
